refactor: 컴포넌트 분리 및 참가자 리스트 모달 추가
